Created time measurement tests, fixed numeric accuracy double bug in Convex Hull Algorithm, implemented help and testCount params

+++ src/model/Point.java
@@ -4,19 +4,19 @@ import java.util.Objects;
 
 public class Point {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -37,11 +37,11 @@ public class Point {
         return Objects.hash(x, y);
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
